PostProData struct from messages to database
